+++ packages/config-resolver/src/endpointsConfig/NodeUseDualstackEndpointConfigOptions.ts
@@ -15,6 +15,8 @@ export const CONFIG_USE_DUALSTACK_ENDPOINT = "use_dualstack_endpoint";
 export const DEFAULT_USE_DUALSTACK_ENDPOINT = false;
 
 /**
+ * Don't delete this, used by older clients.
+ * @deprecated replaced by nodeDualstackConfigSelectors in newer clients.
  * @internal
  */
 export const NODE_USE_DUALSTACK_ENDPOINT_CONFIG_OPTIONS: LoadedConfigSelectors<boolean> = {
@@ -22,4 +24,14 @@ export const NODE_USE_DUALSTACK_ENDPOINT_CONFIG_OPTIONS: LoadedConfigSelectors<b
     booleanSelector(env, ENV_USE_DUALSTACK_ENDPOINT, SelectorType.ENV),
   configFileSelector: (profile) => booleanSelector(profile, CONFIG_USE_DUALSTACK_ENDPOINT, SelectorType.CONFIG),
   default: false,
+};
+
+/**
+ * @internal
+ */
+export const nodeDualstackConfigSelectors: LoadedConfigSelectors<boolean | undefined> = {
+  environmentVariableSelector: (env: NodeJS.ProcessEnv) =>
+    booleanSelector(env, ENV_USE_DUALSTACK_ENDPOINT, SelectorType.ENV),
+  configFileSelector: (profile) => booleanSelector(profile, CONFIG_USE_DUALSTACK_ENDPOINT, SelectorType.CONFIG),
+  default: undefined,
 };